requirement.txt
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.5
django-allauth==0.40.0
django-import-export==0.5.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==2.5.14
psycopg2==2.7.4
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.21.0
requests-oauthlib==0.8.0
tablib==0.13.0
urllib3==1.24.3
xlrd==1.2.0
xlwt==1.3.0

